requirements.txt
streamlit
crewai==0.28.8
langchain-google-genai
python-dotenv
setuptools